Move `deck_conflict` module out of legacy core.
