+++ api/tests/opentrons/protocol_api/core/legacy/test_deck.py
@@ -7,7 +7,7 @@ from decoy import Decoy
 from opentrons.hardware_control.modules import ModuleType
 
 from opentrons.protocol_api.core.legacy.module_geometry import ModuleGeometry
-from opentrons.protocol_api.core.legacy import deck_conflict
+from opentrons import deck_conflict
 from opentrons.protocol_api.core.legacy.deck import Deck, DeckItem
